requirement.txt
Pillow==6.2.1
PyAutoGUI==0.9.48
selenium==3.141.0
mss==4.0.3
